Енумы в отдельные классы
