+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelLimit.cs
@@ -162,52 +162,4 @@ namespace Vodovoz.Domain.Fuel
 
 		public virtual string Title => $"Топливный лимит на {Amount} л. №{Id} (Id={LimitId}) Id карты: {CardId}";
 	}
-
-	public enum FuelLimitStatus
-	{
-		[Display(Name = "Не указано")]
-		None,
-		[Display(Name = "Выдан успешно")]
-		Success,
-		[Display(Name = "Отменен")]
-		Canceled,
-		[Display(Name = "Ошибка на стороне сервиса при выдаче")]
-		ServiceError,
-		[Display(Name = "Ошибка на стороне ДВ при выдаче")]
-		LocalError
-	}
-
-	public enum FuelLimitUnit
-	{
-		[Display(Name = "Литр")]
-		Liter,
-		[Display(Name = "Рубль")]
-		Ruble
-	}
-
-	public enum FuelLimitPeriodUnit
-	{
-		[Display(Name = "Разовый")]
-		OneTime = 2,
-		[Display(Name = "Сутки")]
-		Day = 3,
-		[Display(Name = "Неделя")]
-		Week = 4,
-		[Display(Name = "Месяц")]
-		Month = 5,
-		[Display(Name = "Квартал")]
-		Quarter = 6,
-		[Display(Name = "Год")]
-		Year = 7
-	}
-
-	public enum FuelLimitTermType
-	{
-		[Display(Name = "Все дни")]
-		AllDays = 1,
-		[Display(Name = "Рабочие дни")]
-		WorkingDays = 2,
-		[Display(Name = "Выходные дни")]
-		DaysOff = 3,
-	}
 }
