+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelLimitStatus.cs
@@ -1,0 +1,18 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz.Domain.Fuel
+{
+	public enum FuelLimitStatus
+	{
+		[Display(Name = "Не указано")]
+		None,
+		[Display(Name = "Выдан успешно")]
+		Success,
+		[Display(Name = "Отменен")]
+		Canceled,
+		[Display(Name = "Ошибка на стороне сервиса при выдаче")]
+		ServiceError,
+		[Display(Name = "Ошибка на стороне ДВ при выдаче")]
+		LocalError
+	}
+}
