+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelLimitTermType.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz.Domain.Fuel
+{
+	public enum FuelLimitTermType
+	{
+		[Display(Name = "Все дни")]
+		AllDays = 1,
+		[Display(Name = "Рабочие дни")]
+		WorkingDays = 2,
+		[Display(Name = "Выходные дни")]
+		DaysOff = 3,
+	}
+}
