+++ Source/Libraries/Core/Business/VodovozBusiness/Domain/Fuel/FuelLimitUnit.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz.Domain.Fuel
+{
+	public enum FuelLimitUnit
+	{
+		[Display(Name = "Литр")]
+		Liter,
+		[Display(Name = "Рубль")]
+		Ruble
+	}
+}
